Keep the forwarding-method decorator in core.utilities.misc
